bugfix:把神秘数字改为long类型，防止4位不够而溢出

+++ src/main/java/cn/wxxlamp/ApplicationRunner.java
@@ -23,7 +23,7 @@ public class ApplicationRunner {
     public static void main(String[] args) {
         TreasureBoxService treasureBoxService = new TreasureBoxService();
         TreasureBox[] treasureBoxes = treasureBoxService.createTreasureBoxes(BOX_NUM, BEGIN_STR);
-        Integer[] magicNums = treasureBoxService.getMagicNumsFromTreasureBox(treasureBoxes);
+        Long[] magicNums = treasureBoxService.getMagicNumsFromTreasureBox(treasureBoxes);
         treasureBoxService.printMagicNums(magicNums);
     }
 }
